chore: add system documentation
