+++ .vscode/settings.json
@@ -3,12 +3,15 @@
         "AFFILCOUNTRY",
         "coredata",
         "dataframe",
+        "docstrings",
         "dois",
         "Elsevier",
         "levelname",
+        "micromamba",
         "openai",
         "opensearch",
         "pdfplumber",
+        "pyside",
         "pytestqt",
         "qtbot",
         "REFPUBYEAR",
