+++ src/modules/journal_downloader/__init__.py
@@ -1,16 +1,13 @@
 """Defines the module's public interface."""
-from .download_signal import DownloadElsevierThread
-from .downloader import (
-    clear_journals,
+from .constants import ELSEVIER_API, JOURNALS_PATH
+from .download_signal import (
+    DownloadElsevierThread,
+    DownloadJournalsResult,
     download_journal,
-    download_journals,
-    get_journal,
-    get_journals,
 )
 from .journal_sets import load_sets, write_sets
 from .query_signal import QueryElsevierThread
 
-__all__ = ["download_journal", "download_journals",
-           "get_journals", "get_journal",
-           "clear_journals", "QueryElsevierThread", "DownloadElsevierThread",
-           "load_sets", "write_sets"]
+__all__ = ["ELSEVIER_API", "JOURNALS_PATH", "DownloadElsevierThread",
+           "DownloadJournalsResult", "download_journal", "load_sets",
+           "write_sets", "QueryElsevierThread"]
